Time signature events now stored _only_ in reference track of composition. This reference track is now written out to file along with the others.
Haven't yet finished the bit that actually draws the time sigs in the
right places, but most of the layout code is up-to-date.

Extensive mods to pixmaps and sizes in rosegardentransport.ui; this reverts
to the "original" layout, but I reckon changing the layout is probably easier
than fiddling with all the details of size and pixmap.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@1110 cf6a83e7-f017-4305-b0b3-678bde24f230