Rationalised bar/timesig stuff in Segment.h, because I decided it was a bit excessive after all to duplicate all this stuff between Segment and Composition just so as to avoid having a back-pointer from Segment to Composition.  (Well, it wasn't planned to be duplicated; it just happened that way.)
So now Segment has a Composition pointer, which may be zero if it's not in
a Composition (still a legitimate occurrence) but through which it makes
all the timesig/bar/quantizer requests.  Simplifies some things, anyway.

Some work on Mark import from RG2.1, but all I've really managed to do is
make it possible to crash the thing by trying to render files that came
from RG2.1 import.

Sorry, I'm rambling.  Rambling in CVS logs.  Whatever next?


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@1215 cf6a83e7-f017-4305-b0b3-678bde24f230