PianoKeyboard can now "swoosh" - you can hold down your mouse on a note and sweep up and down the keyboard playing all the notes.
Added a new MappedEvent type called MidiNoteOneShot which is
used to send this type of event (finite duration) out to the
SoundDriver when we're not playing.  We could actually do
this with the current (always running) timing loop but an
event this type is pretty useful anyway and it's trivially
supported by ALSA.  aRts implementation as exisiting MidiNote.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk@2163 cf6a83e7-f017-4305-b0b3-678bde24f230