* filled in all the complicated stuff to do with getting simple changes   of routing back from the GUI to the sequencer * some mixer bits & bobs
Things are still less functional on the outside than they were, but they're
a lot more functional on the inside.

Note that the number of audio submasters and audio record ins are now part
of the studio, not part of the configuration.  (They're still in the config
dialog for the moment, but not connected up.)


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk@5190 cf6a83e7-f017-4305-b0b3-678bde24f230