My answer to my: RFE [ 980195 ] tempo from segment to ask how many bars long is segment
* adds a BeatsBarsDialog to ask the user for info about the selected audio segment
* makes use of that info when setting tempo to audio segment duration

NOTE: see the code for comments explaining the bandaid I had to put on to get it out the door.  It's considerably less elegant than I feel like it ought to have been, but I couldn't get the elegant version to work after three days bashing my head against it.  This works.  If it offends you, fix it, and show me what the @#%@#% I was doing wrong the first time around.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk@5743 cf6a83e7-f017-4305-b0b3-678bde24f230