Commit of work in progress because I have a lot tied up in this, and having all the copies in the world in the same physical location is starting to make me paranoid.
None of this new code has been tested yet, but none of it is actually executed
anywhere yet, so there should be no danger of breakage.  Unless I missed
adding a file somewhere.

Much gratitude to Magnus for the immense instrument database all of this revolves
around!!!

Much apology for the biggest commit in history that didn't actually do
anything yet.



git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@7413 cf6a83e7-f017-4305-b0b3-678bde24f230