Simplistic fix for #1838190    Crash in countdown dialog
I'm sure this isn't really a very good solution, but I can only managed to get
my code that prints the FAILSAFE message to trigger with this one rare test
file, in this exact state.  I can't create a new test to reproduce the
circumstances.  The failsafe does prevent the crash in this isolated case, and I
can see no huge harm.

Besides, we don't show the countdown dialog anymore, do we?  Right.  Isn't this
completely redundant?  I just remember what this is.  It used to show time
remaining to the end of the composition, didn't it?  And since we auto extend,
it's just some piece of cruft, isn't it?

OK, that's too much for a commit log.  Any further discussion to the devel list.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@8442 cf6a83e7-f017-4305-b0b3-678bde24f230