merge staff_group_hacking into trunk/
It still has some cosmetic problems in the exported code, but the exported code
works, and this is passing every test I can think to throw at it.  Time to call
it real, to the deafening roar of millions of people clapping with one hand
while a tree falls in the forest where nobody is around to hear it.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@8637 cf6a83e7-f017-4305-b0b3-678bde24f230