* convert kapp*.h to QApplication * fix string conversion error
File still fails to build due to kapp not existing.  I can't figure out where
kapp comes from in the first place.  Inherited from what?  It doesn't seem to be
declared within PeakFile.* at all.  Anyway, I'm not sure what we will ultimately
come up with as far as KApplication vs QApplication and friends and so on (can
we use i18n() or KDE anything inside a QApplication?) so I'm just deferring the
rest of this for later.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/branches/qt4@9154 cf6a83e7-f017-4305-b0b3-678bde24f230