Move AUTHORS info online permanently.
Credit will be given for work directly pertaining to every individual release in
the form of release notes and about box data on a per-release basis, to be
updated for each release to reflect who actually worked on it since the previous
one.

The online page will keep a record of everyone who has ever contributed to any
incarnation of Rosegarden, and the about box will end with a "see link for more
info" that refers to the online page.

(I haven't written the about box yet.)


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/branches/qt4@9841 cf6a83e7-f017-4305-b0b3-678bde24f230