The peanut gallery over here is driving me crazy, so it will be fun to see what the users think.  Pitting our old transport against the one I just committed, everybody chose the old one as being "less realistic, but easier to read"
I edited the huge originals and reproduced the miniatures for this brightened
version, with the saturation on the blue notched down toward white, then put all
three on the screen, and everyone chose the last version over either of the
other two.

Users are fickle.

I figure commit them anyway, because I went to the trouble to make them.  I
actually like the previous version the best too, the one everyone in the peanut
gallery chose out of the three, but in terms of function over form, this
brighter version is probably easier to read, if slightly less aesthetically
pleasing to my eye.

I am finished overcomplicating things for today.  Thank you, and good night!


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/branches/qt4@9956 cf6a83e7-f017-4305-b0b3-678bde24f230