style mines...  I forget what I did since the last commit, but notably I still don't have the combo boxes hovering to suit me yet.  Dammit.  I'm even crawling around in the Qt code itself for hints.
It's annoying how little there is in google land about any of this.  I start to
think I'm probably doing the most compliated Qt4 stylesheet project ever.

NOTE TO SELF: add the don't show again button to the jack dialog that you're
disabling with this commit, because you're too lazy to not commit the file this
time


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/branches/qt4@9988 cf6a83e7-f017-4305-b0b3-678bde24f230