Review in the morning, but I'm hoping it will turn out that this crash I've been chasing all day really was this obvious and stupid.  Whoever added the new QDialogButtonBox stuff did so after the two populate*() calls, and at least one of the crashing widgets in question was being accessed four jumps down in code ultimately called by one of the populate*() calls.  I should have looked at the original 1.7.x code more closely, and I feel silly for taking as long as I did to realize that the crash was in the CTOR, and none of the other possible outcomes I investigated could have actually HAPPENED yet.  Ugh!
The rest of it, the program combo is working for Whysynth, and the synth plugin
layout for "too many controls" is no longer 15,000 pixels tall.  I'm not sure if
I fixed that by commenting out the m_pluginLabel->()hide() call, or if I fixed
that with a previous commit.  I was doing something with not adding things to
the layout if they were supposed to be hidden, I think.  I've totally forgotten
what I was trying to do to solve that problem, but one way or the other,
something seems to have worked.  Just need to chase down the "too many controls"
label not showing up, and that weirdness coming from the sequencer I asked Chris
about.

Sleepy time now.  I should have quit two hours ago.

Oh, and I still have no idea why the slot inside the ifdef was never working
out.  That's probably going to break building without LIBLO.  We could always
just absolutely require LIBLO though.  Why be coy?


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/branches/qt4@10043 cf6a83e7-f017-4305-b0b3-678bde24f230