* Final work on adding/hiding type whatnot and shuffling bits around to improve   the allocation of rows and columns
* Add some stretch to definitely improve the <ports> <id> bit

* reduce vertical spacing of the parameter QGridWidget to 0 to try to cram them
  as tight as possible (improvement, but not as tight as I'd like to see)

* some pretty futile attempts at improving horizontal alignment--either this
  needs a more refined touch than mine, or I'm thinking the only sensible thing
  to do is probably fix SqueezedLabel so it works, and use those, making sure it
  pops up a tooltip with the whole label in it

TODO:

The "editor" button is never enabled for synth plugins that should have editors.
It would be great if someone else could please sort this out.  (Are there any
other plugins that should have editors?  I'm not sure.  Maybe.)

For audio plugins, the category combo trips something in the sequencer that
emits an "(object not found)" error every time you change the combo.  Highly
obnoxious and irritating, although it appears at least superficially harmless.
This really needs to be fixed, and I've asked Chris to look into it if he would
please be so kind, as he will understand the guts of the sequencer and how it
all relates to plugins of various sorts better than anybody currently working
here.

Just rolling the mouse wheel over a combo box can cause random, immediate,
untraceable crashes that appear to be from bad plugins.  We're talking one and
three-line stack traces, with ladspa/fm_osc_1415.so and libpthread.so in them.
This looks bad.  This kind of thing probably used to crash the sequencer on
occasion, and now it crashes the entire application.  Unpleasant.  We should do
something about this if we can, even if it's as lame as assembling a list of bad
plugins we should avoid touching if present.

My widget allocation code is very, very lame.  Though it mostly seems pretty
well good enough for now.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/branches/qt4@10049 cf6a83e7-f017-4305-b0b3-678bde24f230