* faders pretty much sorted now, although I'm still not thrilled with the ticks
* Simplify AddTrackDialog's layout, and fix the width of the offending
  undersized combo

* Audio VU meters get the same sort of surround as the faders now

* Something or other about tweaking height() -1 in the actual VU meter drawing
  code too.  Can't quite remember that one.

TODO: FloatEdit is really quite bad style-wise still.  I have lost a lot of
patience for this kind of thing at this point, so I will probably attack the
stubborn sumbitch with a local stylesheet to cut straight to the chase.  Deep
widgets are so nasty to style, and even the spinbox background pixmap hack
doesn't work here.

I'm contemplating Add Tracks and Change Composition Start and End to add a
groupbox to their layouts for better visual consistency with our other dialogs,
where the controls are in a dark gray pop-out and the buttons are in the
surrounding black.  These are functional, and we probably have enough minor
InputDialog and QInputDialog-based things in here that these aren't serious
standouts, but I'm thinking about it.  Hell, since I already re-implemented
QInputDialog, I could make it put everything in a group box too.  There's the
solution, and the justification at the same time.

But first I must needs pay some attention to SWMBO lest I find myself in the
doghouse.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/branches/qt4@10063 cf6a83e7-f017-4305-b0b3-678bde24f230