Fixed TrackButton instrument select problem.
Replicated the kde3 functionality of passing a number (turns out to be the
position of the instrument in the long list of instruments) with the sub-
menu entry. Here it's passed in the data() parameter of the QAction object.
The data parameter holds a QVariant which is a wrapper class for any of the
usual data types.

Renamed the i counter, count, to match the Studio.cpp implementation and
because I found this functionality very difficult to find.

This code remains somewhat messy (not least, the tabs are horrible!!)


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/branches/qt4@10071 cf6a83e7-f017-4305-b0b3-678bde24f230