* a bit of rearranging in the new device manager, trying to make it more obvious   what things are for by making the "click to connect" list items resemble   buttons
NOTE: there are still some QLineEdit in here I need to style with spot
stylesheets in the .ui file, but otherwise the style is workable, and I have to
say I'm kind of proud of myself for the button thing.  That was a cool idea!

The little left-pointing arrows I made (because I had to do *something* with the
spot reserved for the list branch graphics, and a stack of buttons connected by
a "sprue" looked stupid) suggest to me that if possible it would be seriously
cool to make this work by dragging those "buttons" and dropping them where you
want to make a connection, instead of clicking there and double clicking here.
No idea if that's possible, but if it is, I think it would be JUST the trick.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/branches/qt4@10100 cf6a83e7-f017-4305-b0b3-678bde24f230