* some icon shuffling
* some initial work looking into the broken bank editor

* DISABLE THE AUTOMATIC DEVICE DETECTOR/CREATOR!!  Let it DIE, DIE, DIE!

We need to think of some way to hook "General MIDI Device" in the default
autoload up to something sensible, but we DON'T need to create new devices
beyond that.  If a user has need for more than one playback device, they will be
well aware of that.

See rosegarden-devel for further discussion.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/branches/qt4@10104 cf6a83e7-f017-4305-b0b3-678bde24f230