* tweak this so it actually works...  cheap and ugly, but it works
NB: if we actually go with this, we will DEFINITELY have to explicitly install
the helpers by name, one by one, instead of installing src/helpers/* as is
currently the case.  This would be a wide open door to exploits.  I'm only
leaving it in this completely careless state because all three of these helper
scripts really have to be replaced with something else anyway before any of this
goes public, and so they won't ultimately need to be installed at all.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@10194 cf6a83e7-f017-4305-b0b3-678bde24f230