* fix widespread QSettings amnesia problem
There's apparently some legacy slotOk() that still compiles, but it doesn't
work.  I had fixed this in one case before, but forgot to look for how the fix
might apply elsewhere.  It turns out it was spread out all over the place.

The solution is to replace our ::slotOk() methods with a ::accept() methods, and
then instead of calling accept() at the end of slotOk(), we call
QDialog::accept() at the end of ::accept()

It doesn't look like every instance of this rewrite was strictly necessary, but
I converted them all anyway to keep the idiom the same everywhere, even where it
had its own plumbing such that the slotOK() did in fact work.  In many cases, it
seems we were relying on plumbing probably from KDialogBase to hook some default
connection not specified in our code to slotOk()


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@10254 cf6a83e7-f017-4305-b0b3-678bde24f230