* baby steps on rewriting the project packager in real code (just doing it in   trunk/ as trunk/ is stone cold dead anyway)
(Note that this is slightly less fun than hitting myself in the head with a
brick to see how many blows it takes before I lose consciousness, so I expect a
lot of solitaire games and other compelling distractions before I finally manage
to make this work again.  If I ever quite manage to endure this much pain.
Yerk.)


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@10303 cf6a83e7-f017-4305-b0b3-678bde24f230