* add back a good chunk of missing slots and support functions, primarily from   the Edit menu
The select from start/end, clear selection, and filter selection stuff all
works, even though I kind of made my best guesstimate what to do with some of
this code, and it may need tidying.  Note that the playback pointer is now the
insertion cursor for this kind of thing.  That confused me for a long time, and
I was going to report this as compiling but broken.  Re-testing with a mind
toward the position of the blue cursor, it does actually seem to work.

The toolbar toggle stuff all works, including the buttons on the tools toolbar
or whatever we call the toolbar to toggle all the other toolbars.

The velocity up/down stuff doesn't seem to be working, and the dialog called by
one of those options is at least partially broken.  I expect these will turn out
to be fairly minor problems.



git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@10444 cf6a83e7-f017-4305-b0b3-678bde24f230