Merge of newcontrolrulers branch.
Functional (at the same level that the branch was!)
Only a 'read only' velocity ruler is implemented.
Use the 'Insert Item' selection under 'Control Ruler Actions' menu to turn on the velocity property ruler.
No mechanism to turn it off again.

Here are the details of changes to files outside of the core control ruler set that could (but I don't think will) affect other work.

Panned.cpp - Nothing other than a commented out change to the slotSetPannedRect function (an alternative method for coordinating panner and panned for later discussion).

MatrixWidget.cpp - Code to create and connect the new ControlRulerWidget (a QStackedWidget).

MatrixScene.cpp - Added getCurrentViewSegment method and selectionChanged signal.

MatrixElement.cpp - Keep local copy of velocity (for velocity tool modification) and return it with getElementVelocity.

MatrixView.cpp - Added insert_control_item - not nearly finished yet.


The *ControRuler.* files have had a brutal set of slashing, hacking, cutting, repairing but the structure is now reasonaby solid (especially for PropertyControlRuler).
I've wanted to leave a lot of the old material in until I understand how to re-enable or reimplement so it's still a bit fo a mess at this stage.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@10457 cf6a83e7-f017-4305-b0b3-678bde24f230