* OK, I give up in frustration again.
95% of the time, getCurrentSelection doesn't return properly, and the two IDs
don't match, so nothing happens.  5% of the time, they do match, the signal
fires, and the GUI just hangs.

This is another thing to add to the pile of things I couldn't solve that Chris
is already not getting around to in any kind of hurry, but for my sanity I need
to just say screw it and abandon this now.

I pissed away an entire day trying anyway, which is more than anybody else has
done for this foresaken dialog.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@10524 cf6a83e7-f017-4305-b0b3-678bde24f230