* make (most) file dialogs remember the last path used for that particular   action
(Note this is not "per type used" as specified in the bug report.  We never kept
a separate record of where we got a .mid vs. a .rg &c., but rather where we last
opened, imported, merged hydrogen, &c.  This seems to be suitably similar to our
old behavior without going totally nuts.)


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@10525 cf6a83e7-f017-4305-b0b3-678bde24f230