* minor work on the speedbar (kind of forgotten partway through, as well as unbundling some more data files, all of which will have to sit on the shelf until tomorrow now)
* random formatting and cruft cleanups of little consequence

* main window saves and restores its geometry, with some potentially complicated
things to look out for:

 - Old code to set up a sane default initial size for the main window has been
bypassed and then ultimately removed, after it apparently made no difference
anyway.  It was only half a system for maintaining main window geometry anyway,
and it must have depended on something we no longer get from KDE.  The new
system works splendidly here, and the initial size when I delete all the
geometry bits out of my Rosegarden.conf is always very sensible, and basically
the same as it has always been in Thorn.

 - I would have preferred to use saveState() and restoreState() to snapshot
toolbar positions and other settings in addition to plain window geometry, but
these did not work for whatever reason.  The QByteArray data got written to the
conf file, but restoring it never worked.  I gave up trying to figure out why,
and switched back to save/restoreGeometry() instead.  Fsck it.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@10545 cf6a83e7-f017-4305-b0b3-678bde24f230