Significant progress on the EventControlRuler problem (though still not much to look at!).
Removed SegmentObserver descenct from ControlRuler so that PropertyControlRuler can inherit ViewElementObserver
and ControllerEventsRuler (name to be changed at some point) can inherit SegmentObserver.

New ControlItem, EventControlItem is used to hold the position of the new event items on this ruler.

New ControlTool, ControlPainter works on the control ruler (if you select any tool after turning on the ruler).
Events are not drawn or saved properly yet.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@10575 cf6a83e7-f017-4305-b0b3-678bde24f230