Let's suck it up and admit defeat on the release date.  This requires not only rerunning bootstrap.h but an entire distclean to get the new version to take.
The new target is February 14th 2010, exactly five years after the release of
1.0 on the same date.

We seem to do well with "magical" target dates.

I'm starting to think we should change the working codename to Phoenix too.  But
no, let's not.  Instead, just take a moment to contemplate the irony of that and
move on.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@10641 cf6a83e7-f017-4305-b0b3-678bde24f230