* Replaced the ViewElementList m_selectedElements in ControlRulerWidget with a simple std::vector
Fixes crash on Matrix Editor close.
In my defence, I would like to point out that I inherited use of said inappropriate class.
My crime, if it can be considered as such, was to use one as a data member so that it can be used
to display selected elements correctly when creating a property ruler ...  

git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@10765 cf6a83e7-f017-4305-b0b3-678bde24f230