Three new marks, open, mute/stopped, and harmonic/flageolet!
(Wow, this was a lot of spread everywhere and piddling for three new marks.)

BUGS:

I forgot to include the new test file with this commit.

The glyphs are only mapped for Feta so far, and I did an incomplete job with the
mapping.  Should check the ISO or whatever spec PDF for the correct standard
names for these glyphs.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@10804 cf6a83e7-f017-4305-b0b3-678bde24f230