We do not support adding new data files from a source tree that is not under version control, or which is under "exotic" version control.
If you are adding new data files to a source tree that is not under version
control, then you are doing something wrong.  Adding data files implies working
on development in some way.  If you are doing development work in a tree that is
not under version control, it is very likely that you are wasting your time.

We got a heavily modified copy of some quite old release tarball instead of a patch.
I would have accepted that feature as a patch, but instead, the whole thing was
just wasted.  This kind of waste is to be discouraged whenever possible, and
this safety mechanism is one small way to provide such discouragement.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@10846 cf6a83e7-f017-4305-b0b3-678bde24f230