New zoom panel for the Matrix panner
The "primary" zoom wheel works.  It's set to stop after 20 clicks in either
direction, which seems to be a reasonable limit.  It could go a little higher or
lower, maybe 25.  We should probaby rewire things a bit to keep the
mouse-wheel-over-panner zoom in sync with the zoom wheel, so they behave as one,
in order to avoid sync problems when switching from one method to the other.

The horizontal zoom wheel doesn't do anything yet, and the vertical one doesn't
even have code to connect to yet.  The reset button doesn't work yet.

The layout is provisional.  I think the basic concept is really plausible, and
once I hit on this idea, I went ahead and dove into independent axis zoom after
all.  Damn the torpedoes and all that.  I'll make it all work.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@10942 cf6a83e7-f017-4305-b0b3-678bde24f230