The piano keyboard gets mangled all kinds of different ways, but otherwise this works surprisingly well for one day of work.  I'm going to give up and sneak into bed just before the family wakes up.
Chris, you might want to look over my totally simplistic passing the thumb wheel
value straight over as a scale factor.  Maybe it needs some tweaking to put it
more in line with how the zoom slider used to work (still works for a little
while longer) and maybe it's OK like this.  It increments very gradually instead
of in 10 or so fixed steps, but is this a bad thing?  It feels like the original
existing zoom to me, and seems about right, but it was so simplistic, I worry
about what went over my head.

Tomorrowish:

remove the zoom slider and associated logic (even though it still works)

correct the thick/thin selection rectangle scaling problem that precipitated all
of this work (yes, I guess we'll be doing that after all, eh?)

try to iron out the piano keyboard scaling problems if I can grasp what's going
on (if not, I'm hoping having gotten this far, I might get a nudge from Chris)

Good morning, Europe.  It is ridiculously well past when I should have given up
for the day.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@10943 cf6a83e7-f017-4305-b0b3-678bde24f230