Final draft on fundamental behavior.  Try it.
(With the exception that horizontal and vertical zoom should probably be allowed
to go negative, and they can't.  I half understand what's required to make that
possible, and if I have to code it, I'm really in favor of leaving negative zoom
to the primary/panner zoom and brushing the dust off my hands and calling it a
day.  After all, I began this whole quest with the inclination to abolish
independent-axis zooming in the first place, and this is a limitation I could
live with easily.)



git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@10951 cf6a83e7-f017-4305-b0b3-678bde24f230