An experiment at making the independent zoom modes a little easier to work with.
If a wheel belongs to the "other" zoom mechanism it dims slightly in a way
that's supposed to avoid looking disabled, but merely "less active."  This
provides a visual indication which zoom mode you're currently in, so it's easier
to avoid switching modes unexpectedly.

Ideal?  Probably not.  Workable?  I think so.  I think it's about time to plop
this in the notation editor and call this sub-project finished.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@10952 cf6a83e7-f017-4305-b0b3-678bde24f230