Hi
I tried to fix drag and drop in the PlaylistView.
After some hours, still no success unfortunately.
I'm commiting, what I have now.

The problem is:
I can't make the PlaylistView to use the right mime type.
It uses "application/x-qabstractitemmodeldatalist" instead.

The function that was supposed to set the mime type is not being called.
Suggestions are welcome. Emru





git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11044 cf6a83e7-f017-4305-b0b3-678bde24f230