Try to sort out remaining confusion with the resource system by automating the process of detecting changes and rebuilding data/data.o as required.
* track data.qrc, trigger an update if this file changed (as when some other
  developer addes new resources and runs 'make qrc' to reference them, then
  commits the result)

* track the data files referenced by data.qrc, trigger an update if one of them
  changes (as when making changes to an existing file, like data/rc/matrix.rc)

* 'make qrc' no longer removes any files, leaving it to the other bits of the
  build system to detect when this is required, and act accordingly

This seems to work fairly well here, though it will be interesting to give it a
real test by seeing what happens the next time some developer other than myself
makes changes to some of these files.  This was all prompted when Chris made a
number of changes to the .rc files and even a 'make clean' was not sufficient to
trigger an automatic rebuild of the resource bundle.

I tested it by making random cut and paste changes to data.qrc, using touch on
existing data files to change their timestamps, and manually rebuilding data.qrc
with 'make qrc' to ensure that it triggered a new resource build in all these
cases.  It also trips every time you change any random source file, which is
something I couldn't figure out any good way to avoid, because I wasn't able to
figure out how the "dependencies" target works in order to do something similar
but separate.  Oh well.  In the field, it seems this false positive rebuild
doesn't actually eat that much time, and the only time a real rebuild happens is
when it's required.  At least so far as I can see here.

Feel free to hack on this and improve it.  I suck at Makefiles.  This syntax was
obviously conceived by someone on a PCP trip.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11088 cf6a83e7-f017-4305-b0b3-678bde24f230