Fix shortcuts ABCDEFG for languages that use Latin-style do re mi note names by distinguishing between C the key and C the note.
This also reveals to me that while we have keyboard shortcuts translatable, it
isn't actually possible to translate them as far as I can figure.  My theory is
we need to doctor one of the extractor scripts to pull them out of the .rc files
and put in QObject::tr() calls with the shortcut strings in them.  I'll try that
directly.  If it works, it means we have a way to make life easier for QWERTZ
etc. users.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11159 cf6a83e7-f017-4305-b0b3-678bde24f230