Quick, barely tested hack to fix my previous oversight when changing the installed binary from /usr/bin/rosegarden to /usr/bin/rosegarden-10.02 (for example.)  This rewrites our rosegarden.desktop file into, eg. rosegarden-10.02.desktop and edits the file to point to the correct binary.
I'm debating whether to take all of this all the way and append $(VERSION) to
everything we install.  It seems like it would probably be best, but we haven't
changed our icon in years, and there's minimal room for conflict in the MIME
department, so I don't think it's worth the effort to try to promote 100%
complete separation.  (Especially since this will only be used to keep the new
Rosegarden in parallel with Classic for a comparatively short time, and
afterwards it will just be for maintaining parallel installs of the new version
and its successors, which will all be totally compatible with respect to what's
installed to the system. If that should ever change, we can address it then.)


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11276 cf6a83e7-f017-4305-b0b3-678bde24f230