Constrain Matrix and Notation views of segments to the portion that's inside the boundaries defined by the composition, using Segment::getClippedStartTime()
This behavior turns out to date back to Classic, and it's debatable.  If I had
segment from -2 to 100, composition boundaries at 1 and 100, I could open the
segment, rewind to bar 1, and play only the part of the segment visible on the
segment canvas, while I had the option to move the pointer back to -2 and play
those "hidden" events as well.  It was sort of workable that way, but I think
it's better not to show anything that isn't on the inside of the composition
boundaries.  Those events wouldn't normally play unless you happened to have a
recording count-in, and I don't feel they're part of the normal part of the
composition.

OTOH, if you do have events before the composition start marker, perhaps by
accident, and you have a recording count-in, you can wind up with mysterious
notes that you only hear while recording.

I'm voting with code to take a chance on the latter problem confusing someone,
because it's certain that having all of this mysterious stuff show up in the
edit view confused me.  (Probably in part due to the way we scroll all the way
left on first opening now, which makes it really obvious when you scroll five
bars to the left of where everything starts.) There's really no crystal clear
best solution to this one, and I could see having to answer FAQs either way I
go.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11286 cf6a83e7-f017-4305-b0b3-678bde24f230