two down
NOTE: I tested all of this in Classic to get some sense of what it was for, and
what it should do.  It's interesting to note that in the same quick tests with
the same data, this new one turned out to be reliable, while the old one was
spotty.

Just don't try this with a debug stream visible in a terminal, because the
longer your selection gets, the more ridiculously out of hand the lag gets.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11338 cf6a83e7-f017-4305-b0b3-678bde24f230