Fix #2909193
Don't try to setPlausibleConnection if the studio is utterly empty.

After this change, I was able to load the offending file attached to that bug
report, and then create a device and hook it up.  This seems to fix the crash.

Behavior wise, if I go and assign track 1 to "New Device #1" then all the tracks
appear to  assigned to "New Device" as well, as shown by the IPB, but their LEDs
remain white until I assign them each to an instrument manually.  That looks
like it may be more than a cosmetic problem, but anyway, at least you can
apparently fix it from there even if there is a real problem.  I'm not going to
investigate further at this time.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11374 cf6a83e7-f017-4305-b0b3-678bde24f230