Fix settings amnesia on notation fonts.
This one was really tricky!  m_sansFont (eg.) is a FontRequester

FontRequester is a QWidget, and so we were storing QWidget::font(), the font
m_sansFont was drawing itself with, and not FontRequester::getFont(), the font
the user had chosen using our widgetry.

Amazingly enough, this bug also existed in stable_1_7!



git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11392 cf6a83e7-f017-4305-b0b3-678bde24f230