* add key change function to matrix
* change matrix "Events" menu back to "Adjust" to match notation

* audit the &amp; to make them consistent between views (may still be missing
  some hold-outs, as I wasn't super thorough, but it's improved)

* add open in Matrix/Percussion Matrix to the matrix, controlled by action
  states, so from the one kind of matrix you have the option to open the segment(s)
  in the other kind of matrix

* use menu name "Segment" throughout, not "Segments"

NOTE: I was in the process of hooking up Jog Left/Right and forgot to finish it.
That and the keyboard stuff I was talking about yesterday are the last two
things I know I still need to do for the matrix.

NOTE: I need to explore transposing segments with the matrix pitch highlights,
see how badly it's going to behave, and try to come up with a plan (possibly
just getting the segment transpose and moving the key.getTonicPitch() the
opposite way *may* be all that's required--of course if there are different keys
at the same time in different segments that *aren't* transposed, hooo boy,
that's probably going to be a mess).


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11402 cf6a83e7-f017-4305-b0b3-678bde24f230