Add a getVisibleStaffCount() to NotationScene, since getStaffCount() is in use in at least one other place, and I don't want to bother checking if changing the expected behavior of getStaffCount() would break any other code.
What this is all about is that if you had three segments on the same track and
opened them all in a notation view, it was taking the multi-staff font size
instead of the single staff font size.  This fixes that.  It turns out to be
annoyingly mushy to explain cleanly.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11405 cf6a83e7-f017-4305-b0b3-678bde24f230