I didn't realize I had broken rulers in notation with my new add segment thing.
I fixed it and cleaned the code up to move all of that into setSegments() and
out of the ctor.  No use creating it in the ctor only to delete it immediately
and re-create it segSegments() after all.

This passes a cursory test that velocity, pitch, and reverb are at least doing
something now, but I may have missed something else.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11515 cf6a83e7-f017-4305-b0b3-678bde24f230