I just had to spend a good bit of time explaining the "intuitive" device/instrument model to someone, and it got me thinking about these channel controls.  I already had a "before you touch this, consider changing the instrument instead" tooltip warning people away from the channel control, and I decided to up the ante by disposing of the channel control entirely.
The instrument number is the channel.  This was originally designed as a "loose
correspondence" but I just can't conceive of any possible way the potential
advantage of being able to do this outweighs the confusion this causes.

If there are legacy files where channels and instruments don't match up, they
should still work, with the caveat that in those cases the 1:1 correspondence
will have been broken ahead of time, and the instrument number won't be
indicative of the actual channel used.  It's an edge case I can live with.
Address it in the file loading code if needs be, to enforce the correspondence.

We'll see if anybody screams about this, and if they do, I want real world
examples of why this is necessary, not what ifs and users mights.  Real world.
How does forcing a 1:1 correspondence between instrument and channel ruin your
life as a MIDI user?  I'm betting it just makes Rosegarden less confusing for
everyone.

The only way I can conceive of this ever being a serious problem is in the case
where we had MIDI devices that could comprise fewer than 16 channels, where
users wanted to use a portion of the channels on one MIDI connector for one
piece of (probably '80s legacy) hardware, and the remainder of that spectrum for
another.  If we ever do have devices that can comprise fewer than 16 channels,
we will have to make it possible to define the range of channels they comprise,
so one of them can have #1 to #8 and the another #8 to #16.  This is a big if
anyway, and not currently an issue.  Merely a future design consideration if we
implement an obscure feature that does not yet exist at all.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11524 cf6a83e7-f017-4305-b0b3-678bde24f230