Massage RosegardenDocument so that clearing the command history in newDoc() and in the dtor is optional.
This was necessary because the import studio from file function was wiping out
the entire command history, including the MacroCommand the function itself had
just put on the stack.

The command history was being cleared in RosegardenDocument::newDoc() and again
in the dtor.

That code structure is exactly the same in Classic, and Classic did not have
this problem.  I haven't fully understood the mechanism behind this difference,
but it appears that the old command history mechanism must not have been a
singleton, and that clearing the history of the instance associated with the
temporary document did not wipe out the entire global command history then,
while it clearly does now.

So now, Julie, you can undo "Import Studio" and observe the instant crash that
follows.

Wheeee!!!

I don't know if that crash is somehow related to any of the rest of this or not.
I rather hope not, and don't quite think so, but it's all a bit too nebulous for
my liking.  There hasn't been any good wholesome fun work today at all.
Everything is baffling puzzles and razor wire.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11536 cf6a83e7-f017-4305-b0b3-678bde24f230