Bug Fix: Default Visible controls on IPB now transmit their Controller and Value at start of playback without first manually adjusting the IPB controllers.
NOTE:
* Please test this!

* If this is the expected behavior, it should not take much more effort to make
all visible IPB controllers send as well.  But several functions that add and
remove controllers will need to push the change from the Device to its
Instruments.

* All controllers are sent prior to the program msb/lsb and programchange.
So depending on equipment and the controller, the programchange may overwrite
these initial sends.  Switching this is easy--but is it desirable?

I hope this is what people wanted.

Your feedback is appreciated.



git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11558 cf6a83e7-f017-4305-b0b3-678bde24f230