A mixed day working on progress dialogs.  This is worse than I had it before the last few changes, and I can't figure out what I got wrong.
I'm giving up for the day anyway.

I've implemented the rest of the missing API, and I've tried a new trick to get
the dialog to hover at 100% long enough for you to see it was there before
changing to a new operation.  That was working rather smoothly until I went in
and tried to figure out why "Generating audio previews..." was getting stuck at
0%.  Now that I'm force-feeding it a 100%, I'm seeing the initial progress
dialog come up solid black but for the progress bar itself.  I'm not sure why.



git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11619 cf6a83e7-f017-4305-b0b3-678bde24f230