Well, that seemed to help.
Incidentally, loading the huge MIDI file, progress appears to get stuck at 47%,
which is what that whole operation change timer hack was about.  It still gets
stuck at 47%, but that wasn't the case in a different version I have a patch
for.

That version had an additional hack to try to increment the progress bar
smoothly if it stopped at, say, 47%, and then hover at 100%.  It was plagued
with problems, and I dumped it, but while that hack was in effect, even though
it never worked properly, I saw progress incrementing smoothly to 91% on the
same MIDI file.

I think it's a flooding problem.  Too many updates, too fast.  Nothing gets a
chance to catch up before it's time to go away, and a 500 ms hover at 100% isn't
long enough to let it all catch up.

I wonder if it's worth putting a timer hack in setValue() to make it pause for
100 ms or so before calling setValueReal().  Might be worth a whack to see if it
does anything nice.  The feeding end of all this progress code is doing rather
reasonable stuff, and I think I'd like to see these file load/save progress
dialogs stay, if only they would look reasonable.

I've tested "generating audio previews" in some contexts where PeakFile actually
emitted the signal, so it seems generally OK.  We'll just leave the
setValue(100) hack in AudioFilemManager to take care of killing everything off
promptly when there's nothing to do.  I'm not sure how it used to die off
before.

OK, that's enough of this for one day.  This stupid misadventure swallowed far
more hours than it should have already.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11620 cf6a83e7-f017-4305-b0b3-678bde24f230