Fix nasty progress dialog crashes when dropping audio files into the audio file manager when you have a bad audio path.
My first plan of attack was to intercept close events coming before the minimum
duration had expired, and defer them.  This seems reasonable, so I left the
code, but it didn't cure the problem.

After some other stumbling, it seemed to come down to making slotFreeze() take a
harder approach to shutting everything down.  Maybe I didn't grasp how frozen
slotFreeze() was supposed to leave things.  So now a freeze breaks the minimum
duration timer and it trips the second half of an operation change to flush
everything out.

At one point I also changed the audio file manager dialog to allocate the
progress dialog on the heap instead of the stack, so it would have a lifetime
beyond the scope of that function.  That also cured the crash, but this fix,
leaving the original (reverted) audio file manager dialog is cleaner and more
secure.

Maybe I could have done a block() or something too.  It seems maybe the QTimer
took on a life of its own, and the operation change timer was firing a signal
that was intercepted and attempted to be routed to the no longer existent slot
from the deleted object after it went out of scope.

I'm not entirely certain of the exact mechanics here, but I do feel better for
getting the crash averted without the change to allocation on the heap for this
dialog.  That means I don't have to go edit a million lines of code to change
all the other ProgressDialogs do heap allocation as well, which is a good thing.

And of course I spent my entire day's worth of time on something quite different
from intended, and haven't managed to do anything at all about the rude ways
drag and drop fails in this situation.  I guess at this late date some kind of
interactive prompting with more text for translation is out anyway, but at least
we'll go try to create the directory if it doesn't exist.  That will work in a
majority of cases, and we'll have to come up with something a little more solid
later on.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11626 cf6a83e7-f017-4305-b0b3-678bde24f230