- Test that the audio path exists, try to create if not, and inform the user via   the warning widget (rather than some irritating popup) if successful
- Warn the user (not actually tested) if the attempt at creation failed

- Spend shockingly more time than I set out to invest revamping the warning
  widget:

  - bring the warning category into the message that's passed around (it was
    already there, and just needed to be crammed into the Message type, which I
    did in a rather lame fashion to minimize changes to the existing code)

  - add a new info button

  - do some switcheroo to set the tab label text to something per category

  - set the tab icon per category (although I may get bored/motivated to supply
    some "broken" versions of these icons, this will do for now)

TODO: We need to try writing a temporary file to the audio path if it exists,
and warn if unable to do so.  I just didn't get that far, as it will take a bit
of fiddling with QSomethingTemporaryFile and whatnot, and I got caught up in all
this other craziness.  (Plus I wasted a lot of time trying to put the audio path
warning initiation stuff in the document.  It didn't get hooked up early enough
to do any good there, and that was probably stupid in retrospect.  Oh well.)

The WarningDialog title "Performance Problems Detected" has various issues.
First, I'm enforcing the "all window titles have 'Rosegarden' in them" rule
pretty strictly, and this breaks it.  Second, the text is never legible
initially, showing as "Performance Problems De[_][ ][x]" anyway.  Third, this
text makes little sense if the only thing in there is a benign informative
message.  I have decided to leave all of that alone, because it isn't serious
enough that it justifies doing even more damage to the shattered string freeze.



git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11634 cf6a83e7-f017-4305-b0b3-678bde24f230