Some abandoned work on getting double-click-to-edit to work in the bank editor. I forgot I was working on this when I went to bed last night.  Double clicking makes the tree widget collapse and expand, and the new slot is called in addition, but the function it calls never actually does anything.
I don't know how to get rid of the default double click behavior we need to
override.

I don't know why editing isn't working.  The API says "if the item is editable"
so I guess these items aren't editable.  I tried to fix that with setFlags() I
think it was, and all I managed to accomplish was to break everything.

This needs to get done soon, but I'm not going to work on it today.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11659 cf6a83e7-f017-4305-b0b3-678bde24f230