Works pretty well for standard controllers.  I drew a crazy zig-zag of interleaved controllers with this thing where I went left to right, right to left, left to right, right to left.  The final event is still doubled, and probably coming one step too soon or something.
Pitch bend is still very, very screwy, but at least this attempt, naive as it
is, no longer floods with 15,000 events in 8 seconds, so it may be broken, but
it's no longer harmful.

I don't like having to init() as a refreseh mechanism after all this, but I
can't work out any other way to get there.  I suspect it's because the route
I've chosen to get through this doesn't really mesh properly with how the rest
of the rulers are supposed to be put together, but it's working, it looks like
clear() should take care of deleting and re-creating everything, so I don't
think this will cause memory leaks.

It would be nice to show a running rubber band of where the line is going to go
when holding shift.  I don't know if I'll manage that or not.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11690 cf6a83e7-f017-4305-b0b3-678bde24f230