This tiny change was buried under a mountain I had to dig through to get here.
It worked though.  If you display a bar containing only controllers and rests in
the notation editor, the barline is no longer drawn in red.  What was happening
was the "where's the end of the stuff that thinks it's in this bar" code was
calculating on these zero-visible-duration events, and the bars were coming up
massively short, because it didn't add up at all.  Taking them out of
consideration let's everything else do the job as originally intended.

This bug was present in Classic, and probably all the way back to forever.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11717 cf6a83e7-f017-4305-b0b3-678bde24f230