BankEditor->Import now allows KeyMappings to import correctly.
The bug was in MidiDevice()  in most cases m_keyMappaingList was not being
set in the constructor.

This manifested itself a few revisions back when Chris C used a
new MidiDevice(MidiDevice) statement to replace the old code that just
pointed to an existing device in the temporary RG file.

Solved one bug, and uncovered another.  He didn't create since this bug, it
was already there.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11735 cf6a83e7-f017-4305-b0b3-678bde24f230