Get the stupid progress dialog out of the way of the text codec chooser.
This is driving me quite mad.  slotFreeze() doesn't work, because it won't
hide().  I resorted to many layers of increasingly brutal hackery (not
committed) to try to get around that, and nothing ever worked.  isVisible()
tests FALSE, but it's STILL VISIBLE, and when I tapped into showEvent() and
show() to try to catch what must be sending a follow-up after the isVisible()
test, that code never fired.  It just looks like it thinks it's not visible when
it is.

I give up.  I just had slotThaw() setModal(false) so at least the idiotic
progress dialog could be covered up, and it wouldn't cause serious harm.

Any ideas on this one?  I'm obviously still missing the forest for the trees
somewhere, and I'm sick of this hateful code. It's not very easy to replace it
all with an empty implementation, and I couldn't even get a hack to never have
the thing show itself work.  It would take precious little to convince me of the
efficacy of hunting down all ProgressDialog and CurrentProgressDialog code and
deleting it permanently, never looking back.

Precious little.



git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11751 cf6a83e7-f017-4305-b0b3-678bde24f230