Correct translations for these two as a proof of concept.
We need to hand edit the .ts files to remove the &amp; and the type="obsolete"
for the Classical, Cross, Mensural and Triangle translations, then delete the
redundant set of these translations that are untranslated and have no &amp.
This way we spare the translators having to deal with our screw-up on these
strings.

What happened was the note styles are supposed to be generated instead of hard
coded, but for some reason we hard coded them as a stopgap measure, and someone
put strings like C&amp;ross in the menu.  The code setting up the default note
style configuration page stuff uses the correct generated on demand method
instead of hard coding the style names, so it was generating strings like
"Cross" to try to pair up with "C&amp;ross" and this is why the translation
failed.  (In addition to the other reason why the translation failed, which was
using tr() instead of QObject::tr() on strings that are found in the QObject
translation context.)

The proper way to fix this would be to fix the menu itself, but who really
cares.  Sorry Chris.  The next most proper way to fix this would be to go edit
one of the Perl scripts and add more hard coded strings for it to generate into
one of the .cpp files.  This would give translators "Cross" and "C&amp;ross" to
translate.

Since we're so close to a release and so many people thought they were finished,
what we're going to do is just forget about the damn &amp; in these menu
strings, and use "Cross" on the menu.  This will allow the string to match up
with the other "Cross" when that combo box is populated, and everything will
appear to be working perfectly fine, even though there's still some stopgap hack
going on under the hood (and nobody really cares, because nobody has ever tried
to extend the note style mechanism in 10 years).

I'll finish the rest of these tomorrow.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11765 cf6a83e7-f017-4305-b0b3-678bde24f230