Removed sending bank and program changes when track selected.
Oracle testing with RG classic reveals that banks and programs where not
sent when changing tracks if the there was no change in the check boxes.

The behavior in Rg classic that did transmit these these signals I believe
to be long standing bug that caused erratic midi signals to be sent when
tracks were changed.




git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11928 cf6a83e7-f017-4305-b0b3-678bde24f230