Fixing typo in eventlist.rc to reactivate paste action in EventView.
Enabling context menu in EventView, using a QSignalMapper and the Qt4 custom context menu functionality to recreate what I assume was the previous behaviour.

Re-enabling double click on items in TriggeredSegmentManager to bring up triggered segment in event view by correcting signal name. Also did the same for single click, although this connects to an empty function. Re-enabled the Add button on the dialog as well.

git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11961 cf6a83e7-f017-4305-b0b3-678bde24f230