If the (nasty!) user specifies a -style option on the command line, this behaves the same as if they had turned off the option on the config page.
Ideally we'd keep track of the override state and revert to the previously
stored state upon removal of the override condition, or we'd at least throw a
dialog informing the user that the config has been changed, but I discarded the
former as more of a bother than it was worth to juggle, and the latter would be
a great heap of arcane garbage that makes my translator head hurt.  Plus it's
the sort of thing I would throw on the warning widget instead of in a pop-up,
and I can't be bothered to wire it up so that this text can get to the warning
widget.

I decided not to bother with any of the above, so be aware that specifying a
-style option will turn your Thorn style off, and therefore you probably
shouldn't do that just to test whether this works.  I already did, and it does.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@11994 cf6a83e7-f017-4305-b0b3-678bde24f230