Moving Segments to right past end of composition are no longer trucating
Several issues had to be detangled to get this to function correctly

* Main issue was use of getEndMarkerTime which would report composition end time
instead of actual segmetn end time.

* Secondary issue related to knowing the end time of the segment with the
latest end time of those being moved and determining if the composition needed
to be adjusted.

...

Also mostly resolved auto expand issue when moving segments and view scrolled
to maximum value.  Still a possibility that the composition may auto expand,
but in most caes it will now not auto expand except when expected.

NOTE: Notation, Matrix, and RG Canvas Measure numbering is not in sync...
mostly due to updates related to matrix view and notation view when segment
end creates a partial measure paslt the last visible bar.

Somebody can fix this if they want.  Also note that RG canvas looks wierd when
composition length does not fall on measure line (auto expanding sometimes
creates composition lengths that are not multiples of full bars.




git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@12040 cf6a83e7-f017-4305-b0b3-678bde24f230