Restore soft truncate when resized segment is shorted from right then expanded
* Resized segments now mimic RG classic when not near the end of composition.
This is the basic functionality that was restored

* Segments that are hidden due to composition end are restored to their upo
their original length when moved to the left.  This is new compared to RG
Classic.

So the beginning of the composition and end of composition behave in mostly
parallel manners.  They both soft truncate the segment (user experience not
the implemtation).

The only know issue is noted in the commit files.  That is when a resized
segment is moved at or passed the end of the composition.  In this case
moving the segment to the left with the select tool or move tool triggers
the resizing upto its original full length.



git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@12057 cf6a83e7-f017-4305-b0b3-678bde24f230