EXPERIMENT: There are three very old bugs involving tied notes and the matrix. These problems are really tricky to solve well, because the properties in question belong to two different events.  If you want to keep them linked, you have to see that you've got part of a pair in your net, and go hunting for the missing piece, if it even exists.  I've thought about that for years, and never made one step forward.  Nobody else ever has either.
So how about if we're never going to make this function nicely (which is fair to
say after this many years), we at least provide a tool for dealing with the
situation?

In this first stab at it, tied notes are shown on the matrix using a different
fill patern, to set them apart from normal notes.  You can still do bad things,
but at least you have a clue that there is something different about these
events.  They also carry a tool tip; my thinking being the user might go "Why is
this note different?" and hover over it long enough for the tip to pop up.

It's miles away from an ideal solution, but maybe it's workable.  If so, we
should do the same kind of thing in notation too, so if you've accidentally
broken up a tied set (which happens extremely frequently) there is a visual clue
that there are broken tied properties.

Thoughts?


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@12121 cf6a83e7-f017-4305-b0b3-678bde24f230