I feel like committing a minor victory, and doing it directly in trunk/
This is very preliminary, and I have a long way to go, but I finally figured out
a route to dealing with the tied-notes-in-the-matrix problem that's looking
viable.  This only works under extremely limited circumstances at the moment,
but it does show promise.  Under lab conditions, click a tied note in the
matrix, and watch all of that note's companions get pulled in along with it.

I wanted to enforce harder links on the ties at a lower level than this, but
doing it at this fairly high level is much less complex, and has finally
produced a result.

Where this is going is to work on the same kind of enforcement tactics in many
other locations in order to produce the same kind of selection behavior
everywhere.  This will also benefit the notation editor, as the algorithm that's
bred in the matrix gets ported over there too.

After literally years of chewing on this problem off and on, it's pretty
exciting to see a result, and a promising approach at long last.  I finally
thunked up sumpin that looks like it might fly.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@12122 cf6a83e7-f017-4305-b0b3-678bde24f230