This fixes the grace note bug.  I didn't understand why you didn't test for grace notes in XmlStorableEvent too, and just //ed out those lines, so I tried to put the grace note test there.  It didn't work, so your solution seemed the most fitting thing to do, and I wound up //ing out the lines in my version too.
There could well be ramifications somewhere else.  I'm not very comfortable with
this solution in terms of feeling secure that I understand all the potential
side effects.

Nevertheless, it does fix the immediate problem with grace notes getting
destroyed by saving and reloading a file.  That was a lot more serious than the
zero-duration notes problem we were trying to fix in the first place.

I haven't been able to load old files and reproduce them as intended, but when
creating new files with grace notes, saving, and reloading them, they're coming
out the same way on both sides.  We might need to try to catch corrupted grace
notes in the file load code and repair the damage.  What is the damage exactly?
That they even have a "duration=" line in the XML at all, I think is what I'm
seeing here.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@12148 cf6a83e7-f017-4305-b0b3-678bde24f230