Add configuration page option to control whether Rosegarden establishes default JACK connections, per request of some group for whom all of this is a serious problem.  This code was NOT TESTED in any way, because I can't get JACK running without interrupting some work I have in progress elsewhere.  I'm going to leave it to the affected users to test this.
This commit also includes some exploratory work on a LilyPond export bug.  We're
exporting empty bars as s64 s1 instead of just s1.  I have a feeling this is
another unintended consequence of the effort to abolish zero-duration notes.
What used to be 0 is now coming out 64 and that's screwing things up.  The
deeper question is why they're showing up in the first place, but I'd settle for
just putting things back like they were and looking away, which I shall endeavor
to achieve forthwith.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@12153 cf6a83e7-f017-4305-b0b3-678bde24f230