De-fuzz most of the notation icons.  At some point Qt stopped being able to cope with the extra "spill over" pixels I was taking advantage of when I did these originally.  The icons have to be no more than 22 pixels in any dimension, or Qt will scale them to that size, with occasionally grizzly results.
I didn't redo the main note and rest icons, because these went up to 24x24, all
of them were at least x24 in one dimension or the other, and a lot of them
wouldn't have fit in x22 without physically rescaling them, which caused exactly
the problem I was trying to avoid with all of this.  I think they actually look
fine for the most part, except for the ones with huge numbers of flags.  Qt is
probably rescaling all of these, but the stems are coming out as a crisp single
pixel in spite of this, and I've decided to leave well enough alone instead of
redrawing all of them with fonts one point smaller.

I didn't bother doing anything with the step-by-step icon either, because while
it's being rescaled, it still looks satisfactory to me.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@12158 cf6a83e7-f017-4305-b0b3-678bde24f230