Phase 1 of the rewrite.  This is promising, and makes me feel really stupid for misreading the comment in Selection.h that caused me to think I would have to go to the ridiculous lengths that were seen in my first attempt at this.
BUGS:

The notation editor doesn't reflect additional notes in the selection,
even though they're part of the selection, and will be cut if the selection is
cut.

Undo that cut, and only the note that was highlighted in blue will be restored.
Everything else is deleted permanently!!!  This may be due to the selection not
expanding the sides of its "box" to accommodate the extra notes that were pulled
in.  (See laborious comment in the code).

Nothing is yet done in removeEvent() or whatever.  We need to have the same
follow-the-links behavior when removing an event too, but I haven't studied that
problem yet.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@12161 cf6a83e7-f017-4305-b0b3-678bde24f230