Restored tied note selection through chords and polyphonic-like structures.
I broke this behavior while fixing more serious select / deselection bug. But
it is now working again.

Test cases fixed:

Setup
* One measure segemment starting on measure 1 (4/4) time
* type: aahnahd8aa
* tie all of middle c notes (keypress a's)

Test 1:
select quarter note on beat or or two

result:
Fail to select tied not in chord or the eighth notes in beat 4

Test 2:

Select either of the eighth notes on beat 4

result:
Fails to select tied notes in beats 1, 2, or 3

Test 3:
Select the tied note in the chord on beat three

result:
falied to select any of the notes it is tied to in beats 1, 2 or 4.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@12195 cf6a83e7-f017-4305-b0b3-678bde24f230