Here's another patch to scratch one of my own itches. It adds a header and implementation file in src/document/io with an abstract LilyPondLanguage class and concrete implementations for every language supported by LilyPond (for note names and accidentals). The LilyPondExport class has been modified to include a pointer to the abstract class, which is initialized in the constructors with a concrete implementation (and deleted in the destructor).
For now, the pointer is initialized with the Nederlands implementation, to be
consistent with the current behavior. A follow-up patch will allow the user to
select the LilyPond language in the export dialog, and save it as a QSetting in
the LilyPondExportConfigGroup (probably as a string that will be passed to a
factory method).

I don't normally work in C++, so any feedback is most welcome.

Submitted:

Daren Beattie ( dtbeattie ) - 2011-03-16 00:40:57 EDT 


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@12267 cf6a83e7-f017-4305-b0b3-678bde24f230