When toggling the solo button, individual edit views were looking at the track where m_segments[0] resided in order to decide what state to display their individual solo button in.
The idea seemed to be that the state of the button could be used as a visual
indication of whether you were about to hear what you were looking at when
hitting play.  If the segment you were editing wasn't on the selected track, and
the transport was in solo mode, the button did not display in a pressed state.

I see the idea behind that, but I don't think it was thought all the way
through, even for the case where an edit view could only display segments from
one track.  Considering that both major editors can display segments from
multiple tracks, there is no chance of doing anything meaningful with this
button state as some kind of indicator.

As such, I altered the behavior so that every edit view will update itself to
reflect the master composition solo state, and the button is in the same state
on the master transport as well as on all the individual transport toolbars.

I think this has probably been about the only logical thing to do with the
button state for some considerable time now.  If people wanted some indication
whether some portion of an individual edit view is going to be heard when they
press play, then we'd have to invent something new to fulfill that purpose, and
it would be quite involved to implement too.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@12294 cf6a83e7-f017-4305-b0b3-678bde24f230