Remove the blank measure at the end of the LilyPond scores exported with repeating segments (with or without volta).
This is a quick hack which breaks some features of the exporter (as
exporting tempo change).

The LilyPondExporter.cpp needs some cleanup and rewrite.
Still some times have wrong values in the exporter and the indentations of
the LilyPond score are broken.

Furthermore, time signature changes may break the repeat with volta and
are not currently handled at all by the LilyPondSegmentsContext class.



git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@12660 cf6a83e7-f017-4305-b0b3-678bde24f230