Get this progress committed before I accidentally revert something...
Gone is CreateAnacrusisCommand, because I ended up just putting together a
MacroCommand out of existing building blocks instead.  Gone is "Add Anacrusis"
from the notation editor, because the notation window onto a certain subset of
segments didn't strike me as a good place to be performing this action after
all.

The new feature is working, but incomplete.  I want to move the initial time and
tempo changes automagically as a final part of this operation, but that code was
a spectacular failure that took me into a twisted alley full of aborts and
segfaults.  I decided to erase every bit of that and start thinking about that
one completely from scratch, because I obviously had a stupid approach.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@12661 cf6a83e7-f017-4305-b0b3-678bde24f230