Introduction of new CompositionObserver::tracksAdded() notification. SequenceManager has been modified to handle this new notification. It is only sent so far by AddTracksCommand.
Since the call to notifyTrackChanged() is still in Composition::addTrack(),
redundant notifications are sent to SequenceManager.  SequenceManager
appears to handle this fine.

Test suggestions: Add some tracks and use the sequencer.  Do some recording
and playback to make sure the new track is working.

Fixed some warnings in SequenceManager.cpp.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@12768 cf6a83e7-f017-4305-b0b3-678bde24f230