Controller 82 use notifyTrackSelectionChanged()
Also fixed a very minor bug with controller 82.  An input of 127 would result
in selection of a track beyond the last, which would be ignored.  This would
cause trouble only in the esoteric situation where you have a switch 
controller that sends 0 or 127.  It would not work.  Thanks, vmpk, for being
the perfect test controller.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@12944 cf6a83e7-f017-4305-b0b3-678bde24f230