Add setting to turn off CompositionView drawing.
In the interests of performance testing, I need to isolate each part
of the system that is using CPU, and CompositionView is one of the
biggest users.  This is a setting that needs to be set manually
in the settings file, so it should have no effect on normal use
of rg.  Intended for debugging only.  Default is "1".  If set to 0,
CompositionView will no longer do any drawing.

See ~/.config/rosegardenmusic/Rosegarden.conf

[Performance_Testing]
CompositionView=1


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13253 cf6a83e7-f017-4305-b0b3-678bde24f230