#include <algorithm> for std::find
Looks like my #include cleanup in r13306 was a bit overzealous.  Although
it compiles fine for me, Tom has reported a problem with finding 
std::find().  This is consistent with my removal of <algorithm> from the
includes.  Why it builds fine for me?  Who knows.  Anyway, doing this
on speculation, hoping this will clear up Tom's build issue.  It
makes sense, so it shouldn't be harmful.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13311 cf6a83e7-f017-4305-b0b3-678bde24f230