Comments.
Some rearranging too.  One slightly interesting rearrangement was the
reordering of the arguments to these two functions:

  * CompositionModelImpl::makeNotationPreviewRects()
  * CompositionModelImpl::makeNotationPreviewRectsMovingSegment()

The original order had the output argument at the beginning of the list.
Not sure what the rg standard is, but since there was a mix of orders
in this file, I decided to go with the order that I think makes the most
sense: inputs then outputs.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13327 cf6a83e7-f017-4305-b0b3-678bde24f230