Fix refresh bug.
When deleting a segment from a track with multiple overlapping segments,
the tracks wouldn't refresh properly.  This was introduced in r13273
and was (thankfully) only present in 13.06.

While signals and slots are a wonderful and powerful way to decouple
objects, they have the serious drawback of subverting C++'s strong
typing.  To the point where one must proceed with extreme paranoia
whenever working with them.  This is not the first time they've burned
me.  And I'm sure it won't be the last.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13367 cf6a83e7-f017-4305-b0b3-678bde24f230