Remove duplicate menu items in Matrix Editor.
This one might be controversial.  We were getting warnings from Qt related
to duplicate menu items.  So, I removed them.  The menu items in question:

  Edit > Insert Pitch &Bend Sequence...

This is already in the Controllers menu.

  Controllers > Paste Events (&V)

This is already in the Edit menu.

If we're really attached to these duplicate menu items, we'll need to find
a way to appease Qt.  Or else we won't be able to benefit from it catching
other errors.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13379 cf6a83e7-f017-4305-b0b3-678bde24f230