SelectDialog coming together piece by piece.
* Adjust the style of pressed QPushButton to enhance contrast
* Implement the use all button for double whole notes

This is impressively huge and complicated to build.  The tabs for marks and
indications will be built along the same lines, and just as complicated, then
there are random other things to sprinkle in here and put somewhere.  This is
taking a lot of time to build, and it's slow going.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13501 cf6a83e7-f017-4305-b0b3-678bde24f230