SelectDialog - compile check and commit progress
Dialog to have four tabs: Duration, Pitch, Special, Advanced

Special tab collected, pruned, organized, and laid out

Advanced tab created for things that wouldn't work on Special

As fun as all this is, the real pain is going to come when I have to write a
function that uses a stupefying collection of tests to decide whether to add an
event.  These things all have to be || together as a practical matter, and my
mind balks at seeing through this far enough to be sure that's going to work out
like I expect in every one of a thousand edge cases.  This is truly an insane
idea!



git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13508 cf6a83e7-f017-4305-b0b3-678bde24f230