Incremental work on SelectDialog
Note that I got off by 1 in here and have a different number of columns from
planned.  I spent forever shuffling around to get to this point before realizing
my mistake, and I'm just going to leave it alone.

I still need the connect() block and all the useRowX slots, and then of course I
have to turn around and go in for round three on my code writer script to
genereate the Advanced tab.  Yeeha.

And then, the pitch chooser, which will take some time for sure.  It's going to
be a gigantic horizontal piano keyboard where every key is checkable, inspired
by but with no code borrowed from VMPK.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13521 cf6a83e7-f017-4305-b0b3-678bde24f230