Disable destructive actions in playback/record
Fix for bug #837.  (More like a fix for what bug #837 evolved into.)

Most of the destructive menu items are now disabled when in record
or playback.  This fixes crashes in a number of use cases.  A few
tricky ones remain.

Also fixed a bug where File > Revert was always disabled.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13676 cf6a83e7-f017-4305-b0b3-678bde24f230