Bring FileSource up to date with Sonic Visualiser
This patch switches FileSource from QFtp and QHttp, which are not available
in Qt5, to the more future-proof QNetworkAccessManager.

Also fixed a bug that broke ftp transfers.

At this point in time, the only way to test this is by dragging and dropping
URLs from a web browser into rg.

The diff is an absolute mess.  I would recommend comparing against
FileSource in Sonic Visualiser instead of trying to make sense out of this
diff.

Thanks to Tim Munro for his patch related to this.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13704 cf6a83e7-f017-4305-b0b3-678bde24f230