Fix crash in DeleteTracksCommand
Caused by the old "deleting the element the iterator is pointing to"
problem.  This one escaped cppcheck because the erase() is done
indirectly via Composition::detachSegment().

This code is actually not used right now as the clients of 
DeleteTracksCommand delete the segments before using 
DeleteTracksCommand.  That results in two commands in the history and 
burdens the user with two undos.  This fix is in preparation for 
addressing that.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13750 cf6a83e7-f017-4305-b0b3-678bde24f230