MIPP: Simplify widget showing/hiding
I've removed the checks to avoid hiding/showing already hidden/shown
controls.  These were probably done for efficiency, but the return
isn't worth it.  Especially given that the first setupForInstrument()
comes in before the first show().  This causes some bugs related to
the fact that a widget that hasn't first been shown is both not visible
and not hidden (!).  Removing the efficiency checks results in much
easier to read code.  Qt's probably smart enough these days to do the
efficient thing anyway.  Why second-guess it?


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13817 cf6a83e7-f017-4305-b0b3-678bde24f230