MIPP: Remove updateBankComboBox() side-effect
Previously, updateBankComboBox() would select the first bank if it
encountered a bank it didn't expect.  This was used to make sure
that the Instrument was valid wrt the Device.

There are two key use cases that are affected by this.  The first
is removing the selected bank from the Device.  Now, the code will
display no selection on the Bank combobox (-1) and preserve the
original bank.  This can be helpful in case the user puts the bank
back in.

The second use case is "Receive external" and an incoming bank that
isn't in the Device.  The same thing will happen here.  The bank
and program change will be preserved, and the Bank combo will show
no selection (-1).  If the user then imports the correct device
file, the combobox will be updated to show the correct bank.

It's not an ideal solution, but it seems no worse than the previous.
Key advantages include simpler code and preservation of bank/program
change when the Device doesn't match.  Key disadvantage is presenting
a "no selection" combobox to the user.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13832 cf6a83e7-f017-4305-b0b3-678bde24f230