MIPP: Remove updateProgramComboBox() side-effect
Since slotSelectBank() no longer allows invalid program changes through,
this side-effect is not needed in the most common use case: selecting
a bank that doesn't happen to have the currently selected program change.

As with r13823, this changes the Program combobox's behavior when faced
with unexpected (not in the Device) program changes due to either bank
editing or "Receive external".  This will now result in no selection (-1)
on the Program combobox and preservation of the unexpected program
change.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13833 cf6a83e7-f017-4305-b0b3-678bde24f230