MIPP: Improve program navigation (variations)
Previously, when in variations mode, the user would have to navigate
back to the base (0th) variation in the Variation combobox before the
other base variation programs would appear in the Program combobox.
Now, the Program combobox always shows all of the base variation
programs, allowing the user to jump directly to any of them.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13838 cf6a83e7-f017-4305-b0b3-678bde24f230