MIPP: Handle unexpected banks
Now, if the MIPP is faced with a bank that isn't in the Device, it
will show the bank widgets to make sure the user can pick something
else if they want to.

Also removed the updateVariationComboBox() side-effect.  This was the
final MIPP widget update side-effect.  Now, the routines in MIPP
either display the Instrument on the widgets, or make changes
to the Instrument based on user input.  They do not make changes
to the Instrument on their own.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13841 cf6a83e7-f017-4305-b0b3-678bde24f230