MIPP: Avoid getting stuck in variations
If the current variation is invalid for a program, the MIPP will now
show the usual unselected Variation, and offer the valid variation(s)
so the user can get out of this situation.

At this point, the MIPP should be able to handle anything thrown at
it.  It will leave unexpected values alone while always providing a
way for the user to select something else.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13842 cf6a83e7-f017-4305-b0b3-678bde24f230