Introduce new Instrument change notification
InstrumentStaticSignals::changed() has been introduced to
eventually replace RosegardenMainWindow::instrumentParametersChanged()
for notifying various objects that an Instrument has changed.
  
To allow for a gradual switchover from the old to the new,
the new mechanism has been connected to the old.  MidiMixerWindow
has been transitioned to the new as a proof of concept.
  
Remaining steps: transition the rest of the receivers, transition
all senders, cleanup.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13859 cf6a83e7-f017-4305-b0b3-678bde24f230