Transition Instrument panels to ISS::changed()
Instead of IPB handling the old
RosegardenMainWindow::instrumentParametersChanged() and delegating it to
AIPP and MIPP, AIPP and MIPP now subscribe on their own to ISS::changed().

With this, all of the receivers of RMW::instrumentParametersChanged()
have been transitioned to the new ISS::changed().  Next step is to
transition the senders.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13871 cf6a83e7-f017-4305-b0b3-678bde24f230