Switch MIPP and IPP from SqueezedLabel to QLabel
Not sure why SqueezedLabel was being used given that setFixedWidth()
was being called on it to disable squeezing.  Regardless, there appears
to be some bug in SqueezedLabel that causes problems for the first
update of the labels on the MIPP.  Switching to QLabel fixes the bug.
Now at first launch, the MIPP shows the instrument and connection
labels.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13881 cf6a83e7-f017-4305-b0b3-678bde24f230