MIPP: Disable comboboxes even in percussion mode
This reverses r11923 from five years ago.  It appears that this change
was made related to saving percussion mode in the .rg file.  That was
then implemented in 11924.  It's not clear how the two were related, so
it appears safe to reverse this.  Certainly from the user's perspective,
these two things are unrelated.

Since the behavior of the checkboxes is the same for percussion and
non-percussion modes, it makes sense to enable/disable the comboboxes
the same way in both modes.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13897 cf6a83e7-f017-4305-b0b3-678bde24f230