MIPP: Fix incorrect connection info at startup
Prior to this change, the MIPP would show "No connection" at rg start even
if a connection had been made.  Changing tracks would reveal the real
connection.  Now the connection info on the MIPP is correct at rg launch.

There may be unforseen issues with this, so I've left the old code in
to make it easier to diagnose future problems.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13902 cf6a83e7-f017-4305-b0b3-678bde24f230