Fix "external controller" MIDI volume bug
Previously, it was not possible to adjust the volume of MIDI tracks
remotely via the external controller port.  This was because
MidiDevice::getIPBControlParameters() was being used to get the
list of valid controller values for the track.  That routine drops
the volume controller along with any controllers that aren't being
displayed (for the benefit of the MIDI Mixer window).

This change switches to using getControlParameters() (no "IPB") which
returns all controllers.  This gives the user the ability to change
everything, including volume, even if it isn't visible on the UI.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13905 cf6a83e7-f017-4305-b0b3-678bde24f230