MIPP: Split setupForInstrument() in two.
setupForInstrument() is now displayInstrument() for external clients,
and updateWidgets() for internal callers.
IPP::setSelectedInstrument() was changed to accommodate this.

This change is a bit difficult to read due to some renaming.
IPP::m_selectedInstrument is now private with access through a new
getter.  The setter is non-trivial, so a getter makes sense in this
case.  This change resulted in a large number of renames in MIPP
and AIPP.

IPP::m_instrumentLabel was being used inconsistently, so it is now
a protected object that is used directly.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13907 cf6a83e7-f017-4305-b0b3-678bde24f230