Use AC_CHECK_LIB() to find JACK
...instead of PKG_CHECK_MODULES().

The latest release of JACK in Debian has a bad Cflags line in
jack.pc.  This causes PKG_CHECK_MODULES() to fail to find JACK.
See Debian bug #773329.  AC_CHECK_LIB() fixes this issue.

The main downside to using AC_CHECK_LIB() is that it doesn't
check the library version.  So, this has to be communicated to
the distros via documentation.  That's already covered in
the README and the wiki.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13910 cf6a83e7-f017-4305-b0b3-678bde24f230