Fix zoom bug introduced by r13912
The procedure:

  * Launch rg
  * Zoom out by one click from 100% to 50%.
  * Scroll to the right
  * Note the vast expanse of empty space.
  
This bug was introduced by r13912 because it would store the "current"
sizes after the two swaps had modified them to no longer be the current
size.

Comments have been added to explain why the position of these two lines
is important to proper functioning of this routine.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13913 cf6a83e7-f017-4305-b0b3-678bde24f230