Fix main window resize crash.
Turns out the call to QWidget::setContentsMargins() in
RosegardenScrollView::updateBottomWidgetGeometry() does nothing other
than sometimes send a resize event that leads to an endless loop.
The call to QAbstractScrollArea::setViewportMargins() in
setBottomFixedWidget() does what we need.  We need to set margins
around the viewport, not around the entire QAbstractScrollArea.

Thanks to Tito Latini for finding this crash and tracking down
setContentsMargins() as the culprit.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13919 cf6a83e7-f017-4305-b0b3-678bde24f230