Remove RosegardenScrollView's frame
The code in CompositionView does not properly take into account
the fact that there might be a frame.  Consequently it will draw
all over one.  Now, the frame has been removed and the drawing
is consistent and predictable.  This allowed the mystery tweaks
in updateBottomWidgetGeometry() to be removed.

Some renames and comments included.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13922 cf6a83e7-f017-4305-b0b3-678bde24f230