Fix TrackButtons update issue.
Removed TrackEditor::paintEvent() which was causing trouble.  It was
making changes to the scrollbars and shifting things around.  Things
that should never be done during a paint event.  Now
TrackEditor::slotCommandExecuted() handles all the necessary changes and
calls update() at the end to fire of paint events.  It's still not
right, but at least the display update issues are fixed for now.

Also changed the type of m_trackButtonScroll to be more precise.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13940 cf6a83e7-f017-4305-b0b3-678bde24f230