Simplification
Changed CV::setSnapGrain(bool fine) to setSnapTime(timeT).  Now it's
obvious what a tool is asking for.  Either SnapToBar or SnapToBeat.

Got rid of CV::isFineGrain().  It was just as easy for the tools
to check for the shift key on their own.  SegmentSplitter was a
little different, so I went with a simpler approach there.  Just
display the full instructions in the status bar at all times.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@13984 cf6a83e7-f017-4305-b0b3-678bde24f230