Fix event scheduling in DSSI synth (bug 1109).
The events are possibly unsorted because we can merge events played in
realtime with recorded events (generally sorted but not immediate).

Before the bug fix, an immediate event could occur after a future event
and if the immediate event (realtime) is a note, it is time shifted and
shortened (also with zero duration).


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@14110 cf6a83e7-f017-4305-b0b3-678bde24f230