Fix #1445. NOTE BEHAVIOR CHANGES:
Incoming MIDI events are ignored by the matrix/notation view if the window is
not active.  This breaks step recording with apps like VMPK as a casualty of
preventing unintentional filling of a forgotten window with globs and globs of
unwanted note events that have to be undone one by one.

No grid setting/pencil not selected prevents step recording from working.
Previously, there was an error dialog meant to warn the user of this condition,
and meant to show only one time.  It used to show for every single event, due to
bungled logic.  I decided to dispose of it entirely, rather than fix the logic,
and just fail silently in this case.  It's pretty obvious to me, at least, what
to go jiggle when nothing happens after I press some keys on my MIDI keyboard.
If this ever becomes a user complaint, I suggest working out a way to just go
pick a grid setting or switch the notation tool automatically.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@14117 cf6a83e7-f017-4305-b0b3-678bde24f230