Fix DSSI synth used with unsuitable channel setup after stop-recording.
We use multiple instances with run_multiple_synths and not MIDI channels.

Received a wrong program change and MIDI cc 10 (pan) with value 100
after stop-recording.  If play, we hear a different sound panned
right; after rewind-to-beginning, the program is correct but remains
panned right and it is necessary to send a manual cc 10 to restore
the original panpot.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@14118 cf6a83e7-f017-4305-b0b3-678bde24f230