MidiFile: Improve conductor track handling
Previously, a conductor track with other events (e.g. text and sysex)
would be added to the composition, and the key signatures would not
be copied to other tracks.
    
Now, a conductor track with other events is added to the composition
(to avoid losing the other events) and the key signatures are copied to
other tracks.
    
As before, a conductor track with nothing but key signatures will be
dropped from the composition.  The key signatures will be copied to
other tracks.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@14227 cf6a83e7-f017-4305-b0b3-678bde24f230