MidiFile: Improve Instrument setup on import
This fixes bug #1439, MIDI Import Modifies Bank Selects.  Now,
all bank selects and program changes are imported as-is.  Once
a track is imported, the events at time zero are checked to see
if any of them can be moved to the Instrument configuration.
    
This has caused a cosmetic regression in the names of Segments.
They can end up as "General MIDI Device General MIDI Device #n".
Looking at this next.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@14234 cf6a83e7-f017-4305-b0b3-678bde24f230