MidiFile: Change track/segment naming
Previously, tracks created from different channels in a MIDI track
would be named "Imported MIDI".  Now, they have the same name as the
MIDI track they came from.
    
Previously the MIDI instrument name was used to name the segments.
Now it is added to the end of the track name in parens.
    
Segments are now named after their tracks.
    
If no names are found, we fall back on the general MIDI program
names.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@14235 cf6a83e7-f017-4305-b0b3-678bde24f230