LilyPondExporter: rework algorithm for exporting groups (beams and tuplets).
The look-ahead function allows to detect the real last-note in a beam group,
even when it's followed by a rest which (erroneously...) has the same group id.

This also fixes the lack of beaming in tuplet groups (previously it was left
to lilypond to sort out, which is wrong when choosing "export beamings"),
and then my autoBeaming off commit made it all unbeamed. Now it works as expected.

These two changes make lilypond export look more like the on-screen rendering.

git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@14374 cf6a83e7-f017-4305-b0b3-678bde24f230