Project Percussion, Phase 1: make the two-bar percussion clef a real thing
NOTES:

It came to light that the mapping for the Inkpen font is totally broken.  Switch
to that font, and nothing works anywhere, at all.  My new glyph is broken too,
obviously.  I didn't make any progress figuring out why, and I didn't try very
hard.  The other fonts are working correctly.

The new clef height on staff rendering behavior hasn't been tested, and is in no
way settled.  Originally, my concept was for a magic super clef that would map
percussion for you, but I realized that drum set notation is fundamentally
polyphonic anyway.  Rosegarden can't deal with polyphony within the same
segment, and a multi-segment approach is called for.  That being the case, the
clef's functionality becomes completely arbitrary and unimportant, as pitches
will ultimately get notated at a particular height through the use of
transposition and/or special event properties (to be worked out in a future
phase).

Various tabs were converted to spaces, and I used a pragma to silence a very
verbose compiler warning relating to the const extern stuff.  It obviously works
like this.

The SegmentSplitByDrumCommand doesn't REMOTELY do anything yet, and is not
intended to compile yet.



git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@14445 cf6a83e7-f017-4305-b0b3-678bde24f230