Add "Allow Reset All Controllers" to preferences
...and remove the old "Send all MIDI Controllers at start of each
playback" option.  This feature was requested by Will on the user
list 2/4/2015 "Unwanted CCs".  The request was to provide the
option to disable sending of "Reset All Controllers" events.  The
"Send all MIDI Controllers..." option did more than that.  This new
option only affects the sending of resets.

This fixes bug #1485 and is related to bug #1480.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@14511 cf6a83e7-f017-4305-b0b3-678bde24f230