Next step in notation editor insert note improvement:
Accidental preview should almost work when preference 
"Accidental in one bar..." is set to "Affect only that bar".

The two other cases still need coding.

The only problem not fixed in "Require cautionnaries in other octaves" mode is
that sometimes accidentals are previewed as cautionary but are not kept when
the note is inserted on the notation score.




git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@14542 cf6a83e7-f017-4305-b0b3-678bde24f230