Remove usage of RosegardenMainWindow::self() in TempoRuler.
In a unittest without a mainwindow, this was a null ptr.

To solve this, I created an EditTempoController class, and moved the
tempo-edition code from the mainwindow there. The ruler calls into
the controller directly. Only editTempos() requires the mainwindow
so the controller emits a signal, which the mainwindow connects to.

This also simplifies the code a bit:
324 insertions(+), 404 deletions(-)

git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@14615 cf6a83e7-f017-4305-b0b3-678bde24f230