Fix AudioPluginDialog bug with new gcc
Previously, the AudioPluginDialog would show no plugin at launch
when a button for a selected plugin was pressed.  Now it works as
it should.
    
This appears to be caused by a change in the way std::string::c_str()
works in the latest gcc.  The fix was to change the type of currentId
from "const char *" to QString.  Now instead of holding on to a
pointer into a std::string, we immediately make a copy.  It's safer.
    
Some code cleanup and renaming is in here as well.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@14649 cf6a83e7-f017-4305-b0b3-678bde24f230