Implement new solo button functionality
This change provides functionality for the solo buttons in the
following windows:

  * Transport (TransportDialog).
  * Notation Editor (NotationView, EditViewBase)
  * Matrix Editor (MatrixView, EditViewBase)

Clicking on the solo button in any of those windows will toggle
solo on the selected track.  In the editors, the track that is
being edited becomes the selected track.

This functionality is slightly different from before.  Previously,
once solo was turned on, changing the selected track would cause
solo to switch to the selected track.  Now, the selected track and
the solo track are decoupled.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@14665 cf6a83e7-f017-4305-b0b3-678bde24f230