Fix Graphics Performance preference issue
Previously, first launch of rg would assume a graphics performance setting
of native (Safe).  Subsequent modifications to preferences would
inadvertently change the setting to raster (Fast).  The reason for this
was a disagreement as to what the default should be when reading this
setting from Rosegarden.conf.

Now the default value is written back out based on the default in main.cpp.
So the default value in GeneralConfigurationPage has no effect.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@14788 cf6a83e7-f017-4305-b0b3-678bde24f230