Fix "all piano" issue with empty connections
The vivaldi op44 example has empty connections.  Previously we
would not send out channel setups in this case.  That would lead
to the composition being played with all piano.  Now we send out
channel setups.

Had to rearrange the initialiseStudio() calls to fix this.  Also
removed the call in SequenceManager which was causing duplicate
channel setups during the autoload at launch.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@14906 cf6a83e7-f017-4305-b0b3-678bde24f230