Fix #1548: Last syllable of lyrics is not copied between two linked segments.
The duration of a text event (as a lyric syllable) is 0.
So when such an event was the last to be inserted, it was placed exactly on the
upper limit of the refresh statuses range and never updated.

To add one to this upper limit when the duration of an event is 0 fixes
the issue.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@15037 cf6a83e7-f017-4305-b0b3-678bde24f230