Fix #1551: No sound from an audiofile when pressing rewind then play.
The cause of the bug was in JackDriver::getNextSliceStart(const RealTime &now):
An attempt to store a negative value in an unsigned int was causing this
method to sometimes return a huge negative value.

Another bug is fixed at the same time: the method was returning the previous
slice start rather than the next when its parameter was negative.



git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@15039 cf6a83e7-f017-4305-b0b3-678bde24f230