Show pointer position on new staff when clicking on a different staff
This is the bug I actually wanted to fix in these classes :)

NotationScene::mousePressEvent calls setCurrentStaff() directly, without
going via the code in the view that was also repainting the pointer pos.
Fixed by connecting to the currentStaffChanged signal instead.

This looks like the enabling/disabling of actions for "adopted segments"
might still be wrong too for the same reason, but I don't know how to
test that.

git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@15382 cf6a83e7-f017-4305-b0b3-678bde24f230