XmlExporter: Use memcpy() instead of strncpy()
This should be safe.  And it should get rid of the warnings from
newer gcc versions.  My only concern is that there may have been
situations where strncpy()'s behavior when faced with NULLs was
needed.  But the code doesn't seem to imply this in any way.

In the end, memcpy() is what is intended here, so it makes sense.
Time will tell whether it works.

See Patch #63.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@15599 cf6a83e7-f017-4305-b0b3-678bde24f230