Remove <volume> and <pan> tags for MIDI
Turns out these were only valid for about six months back in 2003
(from r4101 to r4735).  I removed the writing of these and the reading
as well and bumped the file format point version in case we need to
detect this change in the future.

This fixes volume and pan CCs always going out even if the controllers
were removed from the Device.  Saving and loading the file would bring
them back because of the <volume> and <pan> tags.

[feature-requests:#482]


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@15750 cf6a83e7-f017-4305-b0b3-678bde24f230