Set default expression to 127 in rgd files
Unfortunately, these are binary files, so the diff is pretty useless.
I used a script to do this, and did some sanity checks, so this should
be ok.  Try out your favorite rgd files just to be sure.

For testing, you might need to move
~/.local/share/rosegarden/library someplace safe to see the effect of
these changes.  Then move it back in case you've got edits in there.


git-svn-id: file:///home/ted/devel/rosegarden-svn-backup/code/trunk/rosegarden@15777 cf6a83e7-f017-4305-b0b3-678bde24f230