Fix linking with pipewire-jack
Trying to build rosegarden master on a Fedora 34 machine with the
pipewire version of jack installed
(pipewire-jack-audio-connection-kit-devel-0.3.38-1.fc34.x86_64),
rosegarden fails to find the jack libraries when linking.

It does find it running cmake:

```
-- Checking for module 'jack'
--   Found jack, version 1.9.17
```

but the link fails with

```
[ 94%] Linking CXX shared library ../librosegardenprivate.so
/usr/bin/ld: cannot find -ljack
/usr/bin/ld: cannot find -ljacknet
collect2: error: ld returned 1 exit status
```

The reason is that pipewire-jack puts the jack libraries not in
/usr/lib64 but rather in a pipewire-specific directory.
On my machine, `pkg-config --libs jack` returns

```
-L/usr/lib64/pipewire-0.3/jack -ljack -ljacknet 
```

The rosegarden CMakeFile.txt needs to be extended to propagate
the library directory information to the link.  The changes here
get this working for me.