Change "For Notation" to "Exclude From Printing"
This was decided on the user list to be easier to understand.
I've tried to push the changes as far down into the code in
an attempt to adhere to the "code and UI names must match"
rule.  Still have a little work to do.  Not sure whether I
will fix the .rg xml as well.