cmake: Use find_library for JACK
"[...] things aren't usually done this way in CMake, which is why
target_link_directories didn't exist until recently. The "right"
way to use pkgconfig is to follow it with a find_library call in
order to resolve "lib name jack in /usr/lib64" (as returned by
pkgconfig) into a full path (/usr/lib64/libjack.so) to link against.