Fix bug #1620.
In LilyPond language, the end of a crescendo/decrescendo is marked after the
first note following that crescedo/decrescendo.
When the last note of a crescendo/decrescendo is the last note of a segment,
the wanted note doesn't exist, the end of the indication is not marked and
LilyPond doesn't print any crescendo/decrescendo of the segment.
To fix it, the end of this last crescendo/decrescendo of the segment is marked
after the last note of the segment.

There will still be an issue when the crescendo/decrescedo is limited to the
last note of the segment.
In such a case, this last crescendo/decrescedo is not printed, but the previous
ones are.