Fix crash when unquantizing a selection
Quantizer::unquantize(EventSelection *selection) iterates through the
selection and calls setToTarget() which might delete Events from the
Segment.  That, in turn, deletes Events from the selection which will
invalidate the iterator.  Solution is to use the "Increment Before
Use" idiom to avoid using an invalidated iterator.

Lots of cleanup in here as well.