Fix undo toolbar button refresh issue
While debugging deep within Qt, I found out that the "dirty" region of
the graphicsview's viewport is only 4 pixels wide, because
Panned::showPositionPointer calls viewport()->update(oldRect) which is
a 4 pixels wide rectangle. Normally the graphicsview is repainted via
the scene's update() mechanism (which stores a dirtyRegion in
QGraphicsView itself). I think the heart of the bug is a collision
between the two update mechanisms. The scene thinks it's going to
trigger a repaint of the whole view, but the view has stored only a
tiny update region.

This commit fixes it by using scene()->update() instead.
I have of course reverted the commit that removes the update() call
from MatrixScene::handleEventRemoved).

https://sourceforge.net/p/rosegarden/mailman/message/37663494/