TrackButtons: Remove LEDs from archived tracks
Make it more obvious the tracks are archived which puts
less pressure on the archive color scheme.  This will
become more important when "dark mode" is implemented.

The only annoying situation I was able to create was
arming a track for record then archiving.  It still
records.  Probably need to work that out if someone
complains.