APD: Fix knobs not showing their saved settings
Before this fix, each time you launch the dialog for a plugin it would
show the knobs all in the zero position.  Now it shows the knobs as
they were previously set by the user.

If the plugin is changed, then the ports get reset and the new plugin
shows default values.