examples: Use ./audio for audioPath
This reduces the chances that rg will create a ~/rosegarden directory.
It also fixes bwv-1060-trumpet-duet-excerpt.rg creating a directory
whose name is only a space character in ~.

Downside to this is that if someone adds some audio to an example,
they will likely lose it unless they think to find it in ~/.local
and preserve it someplace else.